changed default color to green
I have no idea why this works, but we could not get our red agent to turn red. However, when we made the default color green, the script worked and turned our red agent red. To prove this, we switched the code back to the original, and we lost our red agent, and then we switched the code back to having the default green, and we regained our red agent.